Add files via upload
This file contains the commands used to perform real data analyses with schizophrenia in the HORNET paper. The file paths to eQTL and schizophrenia summary GWAS data files are omitted. If you want to replicate our results, you must use the data we used in the paper and change these file paths to the locations where you have stored the data we used.